requirements.txt
fastapi
uvicorn
pydantic
PyPDF2
Pillow
